backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
supabase
python-dotenv
pydantic
loguru
psycopg[binary]
python-jose[cryptography]
passlib[argon2]
python-multipart

numpy
pandas
joblib
scikit-learn


python-socketio[asyncio]
python-engineio[asyncio]

